convert the body-df to a matrix in setup_heatmap
ComplexHeatmap uses matrices as input, so the tidy data-frame (body_data)
passed into setup_heatmap has to be converted to a matrix.

Adds and tests a function `as_matrix` for converting a data-frame to a matrix.
This function allows the user to specify a rowname_column which should be
pulled out of the data-frame and used as rownames in the resulting matrix.

Package now imports dplyr / tidyr / magrittr