Set plotting formats using format_heatmap() (#8)
Default formats were defined: no rownames, no column-clustering

User-defined formats were allowed (by modifying args to format_heatmap())

Heatmap function has dozens of arguments and produces a graph that can't
readily be unit tested over. We wanted to break the argument list for Heatmap
up so that we can write different formatting functions and that can be cut into
the calls to make heatmaps.

Here we mocked-out the Heatmap function (in ComplexHeatmap) and added a test to
check that any plotting argument that was set during format_heatmap() is passed
as an argument when Heatmap is ultimately called.

We imported `mockery` to achieve this.