add function `as_body_matrix`
The function converts a tidy data-frame into a matrix where the rows and
columns are indexed by user-selected columns of the data-frame. The values in
the matrix come from a single specified column of the data-frame.

A single test was added.

`setup_heatmap` was refactored to use `as_body_matrix` to convert the input
`body_data` entry.