fixes #14: add clusters/splits in format_heatmap (#16)
Pre-existing clustering of the features in a heatmap can be passed through by
setting split = c(colname1, colname2, ...) in `format_heatmap`. Here, the
colnames must be present in the row_data entry of the heatmap_data object.

That the `split` argument isn't passed directly through to Heatmap is annotated
in the docs for `format_heatmap()`.

Tests:

- single column that is present within `row_data`

- single column that is absent from `row_data` throws an error