Row annotations (#17)
* fixes #9: allow row-annotated heatmaps

New function `annotate_heatmap` (and helper `.append_row_annotations`) for
adding rowAnnotations to a heatmap. User must specifiy columns of `row_data`
to be used as annotation columns.

`plot_heatmap` was modified to respect returning HeatmapLists as well as
Heatmaps. It checks whethre `row_annotations` entry is defined for a
heatmap_data object and appends the row-annotations to the heatmap if they do
exist.

tests:

- `plot_heatmap` checks that output from a row-annotated heatmap is a
HeatmapList

- `annotate_heatmap` fails on invalid (non heatmap_data) input

- `annotate_heatmap` adds a `row_annotation` data-frame to a heatmap_data
object (& some validity checks on the row_data entry).