ensure user can format any row-annotations (#22)
Argument `row_dots` to `annotate_heatmap` should be a list. All entries in this
list are ultimately added to the arguments in a `HeatmapAnnotation()` call that
defines how the row-annotations are formatted. This allows the user to specify
any argument (other than `df`) that they would pass into `HeatmapAnnotation()`
when using the package `ComplexHeatmap`.

new function: `.add_row_dots` that adds any annotation-formatting arguments to
the `heatmap_data` object.

test: user can select annotation columns from `row_data` and the annotations
and formats are passed through to `HeatmapAnnotation`.